upload-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0
httpx>=0.27.0
python-multipart>=0.0.9
slowapi>=0.1.9
itsdangerous>=2.2.0
pydantic-settings>=2.5.0
